Need to introduce a common base test class
